Try removing duplicate code
